[REF-5622] Refactor MLPiper and Add command-line integration tests
